Struct render::front::Renderer

pub struct Renderer {
    // some fields omitted
}

Renderer front-end

Methods

impl Renderer

fn reset(&mut self)

Reset all commands for the command buffer re-usal.

fn as_buffer(&self) -> &ActualCommandBuffer

Get a command buffer to be submitted

fn clone_empty(&self) -> Renderer

Clone the renderer shared data but ignore the commands

fn clear(&mut self, data: ClearData, frame: &Frame)

Clear the Frame as the ClearData specifies.

fn draw<P: Copy + Program>(&mut self, mesh: &Mesh, slice: Slice, frame: &Frame, program: P, state: &DrawState) -> Result<(), DrawError>

Draw slice of mesh into frame, using a program shell, and a given draw state.

fn update_buffer_vec<T: Send>(&mut self, buf: BufferHandle<T>, data: Vec<T>, offset_elements: uint)

Update a buffer with data from a vector.

fn update_buffer_struct<U, T: Blob<U> + Send>(&mut self, buf: BufferHandle<U>, data: T)

Update a buffer with data from a single type.

fn update_texture<T: Send>(&mut self, tex: TextureHandle, img: ImageInfo, data: Vec<T>)

Update the contents of a texture.