Trait render::front::DeviceHelper

pub trait DeviceHelper {
    fn create_renderer(&mut self) -> Renderer;
    fn create_mesh<T: VertexFormat + Send>(&mut self, data: Vec<T>) -> Mesh;
    fn link_program<L, T: ShaderParam<L>>(&mut self, vs_src: ShaderSource, fs_src: ShaderSource) -> Result<UserProgram<L, T>, ProgramError>;
}

Backend extension trait for convenience methods

Required Methods

fn create_renderer(&mut self) -> Renderer

Create a new renderer

fn create_mesh<T: VertexFormat + Send>(&mut self, data: Vec<T>) -> Mesh

Create a new mesh from the given vertex data. Convenience function around create_buffer and Mesh::from.

fn link_program<L, T: ShaderParam<L>>(&mut self, vs_src: ShaderSource, fs_src: ShaderSource) -> Result<UserProgram<L, T>, ProgramError>

Create a simple program given a vertex shader with a fragment one. The return type can not be derived from the arguments, so you need to provide it explicitly:

let program: MyProgram = device.link_program(vs_src, fs_src).unwrap();

Where MyProgram type is an alias to UserProgram<L, T> that is typically generated by the #[shader_param(MyProgram)] macro used on the data struct containing shader parameters.