Module render::mesh

Meshes Mesh loading.

A Mesh describes the geometry of an object. All or part of a Mesh can be drawn with a single draw call. A Mesh consists of a series of vertices, each with a certain amount of user-specified attributes. These attributes are fed into shader programs. The easiest way to create a mesh is to use the #[vertex_format] attribute on a struct, upload them into a Buffer, and then use Mesh::from.

Structs

Attribute
Describes a single attribute of a vertex buffer, including its type, name, etc.

Mesh
Describes geometry to render.

Enums

Slice
Description of a subset of Mesh data to render. We provide a primitive type in a slice because it is how we interpret mesh contents. For example, we can have a Point typed vertex slice to do shape blending, while still rendereing it as an indexed TriangleList.

Traits

VertexFormat
A trait implemented automatically for user vertex structure by `#[vertex_format] attribute