Struct render::shade::UserProgram

pub struct UserProgram<L, T> {
    // some fields omitted
}

A bundle that encapsulates a program and a custom user-provided structure containing the program parameters.

Type parameters:

L - auto-generated structure that has a variable index for every field of T

T - user-provided structure containing actual parameter values

Methods

impl<L, T: ShaderParam<L>> UserProgram<L, T>

fn connect(prog: ProgramHandle) -> Result<UserProgram<L, T>, ParameterError>

Connect a shader program with a parameter structure