Trait render::shade::ShaderParam

pub trait ShaderParam<L> {
    fn create_link(Option<Self>, &ProgramInfo) -> Result<L, ParameterError>;
    fn fill_params(&self, &L, ParamValues);
}

Abstracts the shader parameter structure, generated by the shader_param attribute

Required Methods

fn create_link(Option<Self>, &ProgramInfo) -> Result<L, ParameterError>

Creates a new link, self is passed as a workaround for Rust to not be lost in generics

fn fill_params(&self, &L, ParamValues)

Get all the contained parameter values, using a given link.