Enum render::state::BlendPreset

pub enum BlendPreset {
    BlendAdditive,
    BlendAlpha,
}

Blend function presets for ease of use.

Variants

BlendAdditive

When combining two fragments, add their values together, saturating at 1.0

BlendAlpha

When combining two fragments, add the value of the source times its alpha channel with the value of the destination multiplied by the inverse of the source alpha channel. Has the usual transparency effect: mixes the two colors using a fraction of each one specified by the alpha of the source.