Enum render::target::Plane

pub enum Plane {
    PlaneEmpty,
    PlaneSurface(Surface),
    PlaneTexture(Texture, Level, Option<Layer>),
}

A single buffer that can be bound to a render target.

Variants

PlaneEmpty

No buffer, the results will not be stored.

PlaneSurface

Render to a Surface (corresponds to a renderbuffer in GL).

PlaneTexture

Render to a texture at a specific mipmap level If Layer is set, it is selecting a single 2D slice of a given 3D texture